requirements.txt
aniso8601==2.0.0
astroid==1.6.1
attrs==17.4.0
click==6.7
colorama==0.3.9
Flask==0.12.2
Flask-API==1.0
Flask-Login==0.4.1
Flask-RESTful==0.3.6
flask-restplus==0.10.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pluggy==0.6.0
py==1.5.2
PyJWT==1.5.3
pylint==1.8.2
pytest==3.4.0
pytz==2018.3
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
